Added check to put\delete methods if logged in user is owner.

+++ myTrip/photo/views.py
@@ -11,7 +11,6 @@ from trip.models import Trip
 from .models import Photo
 
 
-
 class PhotoView(View):
     """Class that handle HTTP requests."""
 
@@ -47,15 +46,19 @@ class PhotoView(View):
         update_data = json.loads(request.body.decode('utf-8'))
         photo = Photo.get_by_id(photo_id)
         if not photo:
-            return HttpResponse(status=404)
-        photo.update(**update_data)
-        data = photo.to_dict()
-        return JsonResponse(data, status=200)
+            return HttpResponse(status=400)
+        if photo.user.id == request.user.id:
+            photo.update(**update_data)
+            data = photo.to_dict()
+            return JsonResponse(data, status=200)
+        return HttpResponse(status=403)
 
     def delete(self, request, photo_id=None):  # pylint: disable=unused-argument,no-self-use
-        """DELETE request handler.If photo were found by id, try to delete photo."""
+        """DELETE request handler. If photo were found by id, try to delete photo."""
         photo = Photo.get_by_id(photo_id)
         if not photo:
-            return HttpResponse(status=404)
-        photo.delete()
-        return HttpResponse(status=200)
+            return HttpResponse(status=400)
+        if photo.user.id == request.user.id:
+            photo.delete()
+            return HttpResponse(status=200)
+        return HttpResponse(status=403)
